+++ lang/en/theme_boost_campus.php
@@ -81,3 +81,9 @@ $string['footerhidehomelinksetting'] = 'Hide link to the webpage start';
 
 $string['courselayoutheadingsetting'] = 'Course layout';
 $string['footerlayoutheadingsetting'] = 'Footer layout';
+
+$string['blockdesignheadingsetting'] = 'Blocks';
+$string['blockiconsetting'] = 'Block icon';
+$string['blockiconsetting_desc'] = 'With this setting you can add a default Font Awesome icon in front of the block title.  If checked, we additionally provide individual icon replacements for many Moodle core blocks and also some widely used blocks. You also can change the icons easily for each block individually in your raw SCSS via the change of the Font Awesome content. For all available icons please visit http://fontawesome.io/icons/ and use the Unicode value of the icon to replace the default one. <br/> The code to change the icon looks like this example change for the block "People": <br/>';
+$string['blockiconsetting_desc_code'] = '.block_people .card-block .card-title::before { content: \'\f0c0\' ; }';
+